all credits changed to double, not decimal

+++ createTables.sql
@@ -4,8 +4,8 @@ CREATE TABLE Class(
 class_id INT AUTO_INCREMENT PRIMARY KEY,
 title VARCHAR(128) NOT NULL,
 link VARCHAR(2000),
-credits DECIMAL NOT NULL,
-pep_credits DECIMAL,
+credits DOUBLE NOT NULL,
+pep_credits DOUBLE,
 UNIQUE(title) );
 
 CREATE TABLE Student(
@@ -21,8 +21,8 @@ CREATE TABLE Manually_Entered_Class(
 class_id INT AUTO_INCREMENT PRIMARY KEY,
 title VARCHAR(128) NOT NULL,
 link VARCHAR(2000),
-credits DECIMAL NOT NULL,
-pep_credits DECIMAL );
+credits DOUBLE NOT NULL,
+pep_credits DOUBLE );
 
 ALTER TABLE Manually_Entered_Class
 AUTO_INCREMENT = 1000;
@@ -31,7 +31,7 @@ CREATE TABLE Requirements(
 r_id INT AUTO_INCREMENT PRIMARY KEY,
 specialization VARCHAR(128) NOT NULL,
 description VARCHAR(1000) NOT NULL,
-credits DECIMAL );
+credits DOUBLE );
 
 CREATE TABLE Takes(
 user_id INT,
